requirements.txt
numpy
torch
transformers
tqdm
scipy
google-genai
python-dotenv